==> 05-reliability.Rmd <==
---
title: "CDS ODS reliability"
date: "Dec 17, 2022"
output:
  html_document: 
   toc: true
   toc_float: true
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}

library(tidyverse)
library(readxl)
library(irr) # kappa
library(psych) # kappa

```



## Read in data
```{r}
coders_ir_rel <- read.csv("./data/coders_ir_rel.csv")

```


## Exclude data: 1) splits [from rater1 coding] and 2) LENA off child for > 50%
```{r}
# exclude data
coders_final <- coders_ir_rel %>% 
  filter(cds_ods_coder1 != "split") %>% 
  filter(LENA_on_child != "no >50%")
  
# view excludes
# splits
df_split <- coders_ir_rel %>% 
  filter(cds_ods_coder1 == "split")

nrow(df_split)
df_split %>% group_by(study) %>% count()

# pct excluded
pct_split <- (nrow(df_split)/nrow(coders_ir_rel))*100
pct_split

# lena less than 50%
df_lena_off_more_50pct <- coders_ir_rel %>% 
  filter(LENA_on_child == "no >50%")

nrow(df_lena_off_more_50pct)
df_lena_off_more_50pct %>% group_by(study) %>% count()

# pct excluded
pct_lenaoff50 <- (nrow(df_lena_off_more_50pct)/nrow(coders_ir_rel))*100
pct_lenaoff50

# total excluded
pct_split + pct_lenaoff50


# table
table(coders_ir_rel$cds_ods_coder1, coders_ir_rel$CDS_pct_of_AWC)
table(coders_ir_rel$cds_ods_coder1, coders_ir_rel$CDS_pct_of_AWC)/ rowSums(table(coders_ir_rel$cds_ods_coder1, coders_ir_rel$CDS_pct_of_AWC))

```



# Reliability

## Final data details
```{r}
# total kept
nrow(coders_final)
100 - (pct_split + pct_lenaoff50)


# number of rows matches each dataset
coders_nrow <- coders_final %>% 
  group_by(study2) %>% 
  count()
coders_nrow


# confusion matrix for all 
table(coders_final$cds_ods_coder1, coders_final$cds_ods_coder2)
table(coders_final$cds_ods_coder1, coders_final$cds_ods_coder2)/
  rowSums(table(coders_final$cds_ods_coder1, coders_final$cds_ods_coder2))


# confusion matrix - WF
coders_final_wf <- coders_final %>% filter(study2 == "WF")

table(coders_final_wf$cds_ods_coder1, coders_final_wf$cds_ods_coder2)
table(coders_final_wf$cds_ods_coder1, coders_final_wf$cds_ods_coder2)/
  rowSums(table(coders_final_wf$cds_ods_coder1, coders_final_wf$cds_ods_coder2))


# confusion matrix - SOT - Outreach
coders_final_sot_outreach <- coders_final %>% filter(study2 == "SOT_Outreach")

table(coders_final_sot_outreach$cds_ods_coder1, coders_final_sot_outreach$cds_ods_coder2)
table(coders_final_sot_outreach$cds_ods_coder1, coders_final_sot_outreach$cds_ods_coder2)/
  rowSums(table(coders_final_sot_outreach$cds_ods_coder1, coders_final_sot_outreach$cds_ods_coder2))



# confusion matrix - SOT - Stanford
coders_final_sot_stanford <- coders_final %>% filter(study2 == "SOT_Stanford")

table(coders_final_sot_stanford$cds_ods_coder1, coders_final_sot_stanford$cds_ods_coder2)
table(coders_final_sot_stanford$cds_ods_coder1, coders_final_sot_stanford$cds_ods_coder2)/
  rowSums(table(coders_final_sot_stanford$cds_ods_coder1, coders_final_sot_stanford$cds_ods_coder2))



# confusion matrix - CONTX - english
coders_final_contx_english <- coders_final %>% filter(study2 == "CONTX_english")

table(coders_final_contx_english$cds_ods_coder1, coders_final_contx_english$cds_ods_coder2)
table(coders_final_contx_english$cds_ods_coder1, coders_final_contx_english$cds_ods_coder2)/
  rowSums(table(coders_final_contx_english$cds_ods_coder1, coders_final_contx_english$cds_ods_coder2))


# confusion matrix - CONTX - spanish
coders_final_contx_spanish <- coders_final %>% filter(study2 == "CONTX_spanish")

table(coders_final_contx_spanish$cds_ods_coder1, coders_final_contx_spanish$cds_ods_coder2)
table(coders_final_contx_spanish$cds_ods_coder1, coders_final_contx_spanish$cds_ods_coder2)/
  rowSums(table(coders_final_contx_spanish$cds_ods_coder1, coders_final_contx_spanish$cds_ods_coder2))



```


## Proportion of matched codes
```{r}
# create match column
coders_final <- coders_final %>% 
  mutate(cds_ods_coder1 = as.numeric(cds_ods_coder1), 
         cds_ods_coder2 = as.numeric(cds_ods_coder2)) %>% 
  mutate(match = ifelse(cds_ods_coder1 == cds_ods_coder2, "match", "mismatch"))

# count of match
coders_final %>% 
  group_by(match) %>% 
  count() %>% 
  mutate(percent = n/nrow(coders_final))


# count of matches vs. mismatch - by STUDY
prop <- coders_final %>% 
  group_by(study2, match) %>% 
  count() %>% 
  group_by(study2) %>% 
  mutate(study2_n = sum(n)) %>% 
  mutate(percent = n/study2_n) %>% 
  mutate(study2 = factor(study2, levels = c("WF", "SOT_Outreach", "SOT_Stanford", "CONTX_english", "CONTX_spanish")))
prop


ggplot(prop, aes(study2, percent, fill = match)) +
  geom_col(position = "dodge")


ggplot(coders_final, aes(AWC_5min, fill = study2)) + 
  geom_histogram() +
  facet_wrap(~ match, ncol = 1)

```




## ALL data
```{r}
coders_final_kappa <- coders_final %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))

kappa2(coders_final_kappa)
```


## no CONTX
```{r}
coders_final_kappa_noCONTX <- coders_final %>% 
  filter(study != "CONTX") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))
  
kappa2(coders_final_kappa_noCONTX)
```


# Reliability - per dataset
## WF only
```{r}
# we do not want weighted kappa; this takes into account disagreements, but weighted kappa is more relevant when data are rank ordered
coders_final_kappa_wf <- coders_final %>% 
  filter(study2 == "WF") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))
  
kappa2(coders_final_kappa_wf)
```


## SOT Outreach only
```{r}
coders_final_kappa_sotOutreach <- coders_final %>% 
  filter(study == "SOT_Outreach") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))
  
kappa2(coders_final_kappa_sotOutreach)
```


## SOT Stanford only
```{r}
coders_final_kappa_sotStanford <- coders_final %>% 
  filter(study == "SOT_Stanford") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))
  
kappa2(coders_final_kappa_sotStanford)
```


## CONTX only
```{r}
coders_final_kappa_CONTX <- coders_final %>% 
  filter(study == "CONTX") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))


kappa2(coders_final_kappa_CONTX)

```


## CONTX only - ENGLISH
```{r}
coders_final_kappa_CONTX_en <- coders_final %>% 
  filter(study2 == "CONTX_english") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))

kappa2(coders_final_kappa_CONTX_en)
```


## CONTX only - SPANISH
```{r}
coders_final_kappa_CONTX_sp <- coders_final %>% 
  filter(study2 == "CONTX_spanish") %>% 
  dplyr::select(c(cds_ods_coder1, cds_ods_coder2))

 
kappa2(coders_final_kappa_CONTX_sp)
```
